Adding

+++ src/main/java/com/decisionTree/wiki/dao/QuestionGroupRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface QuestionGroupRepository extends JpaRepository<QuestionGroupDomain, Long> {
+public interface QuestionGroupRepository extends JpaRepository<QuestionGroupDomain, Integer> {
 }
