+++ src/main/java/com/decisionTree/wiki/dao/TreeRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TreeRepository extends JpaRepository<TreeDomain, Long> {
+public interface TreeRepository extends JpaRepository<TreeDomain, Integer> {
 
     TreeDomain findByRoot (int root );
